Fixed tests for new behaviour of `kill_process`
